Use T4.BuildTools for TheoryData related classes
